refactor: 변수명, 메서드명, 클래스명 변경
